# Conflicts:
#	cmake/compiler/gcc/target.cmake
#	drivers/serial/CMakeLists.txt
#	drivers/serial/Kconfig

+++ scripts/ci/run_ci.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 # Copyright (c) 2017 Linaro Limited
+# Copyright (c) 2018 Intel Corporation
 #
 # SPDX-License-Identifier: Apache-2.0
 #
@@ -15,7 +16,7 @@
 # -r  the remote to rebase on
 #
 # The script can be run locally using for exmaple:
-# ./scripts/ci/run_ci.sh -b master -r upstream  -p
+# ./scripts/ci/run_ci.sh -b master -r origin  -l -R <commit range>
 
 set -xe
 
@@ -25,14 +26,20 @@ SANITYCHECK_OPTIONS_RETRY_2="${SANITYCHECK_OPTIONS} --only-failed --outdir=out-3
 export BSIM_OUT_PATH="${BSIM_OUT_PATH:-/opt/bsim/}"
 export BSIM_COMPONENTS_PATH="${BSIM_OUT_PATH}/components/"
 BSIM_BT_TEST_RESULTS_FILE="./bsim_bt_out/bsim_results.xml"
+WEST_COMMANDS_RESULTS_FILE="./pytest_out/west_commands.xml"
 
 MATRIX_BUILDS=1
 MATRIX=1
 
-while getopts ":p:m:b:r:M:cfs" opt; do
+while getopts ":p:m:b:r:M:cfslR:" opt; do
 	case $opt in
 		c)
 			echo "Execute CI" >&2
+			MAIN_CI=1
+			;;
+		l)
+			echo "Executing script locally" >&2
+			LOCAL_RUN=1
 			MAIN_CI=1
 			;;
 		s)
@@ -63,6 +70,10 @@ while getopts ":p:m:b:r:M:cfs" opt; do
 			echo "Remote: $OPTARG" >&2
 			REMOTE=$OPTARG
 			;;
+		R)
+			echo "Range: $OPTARG" >&2
+			RANGE=$OPTARG
+			;;
 		\?)
 			echo "Invalid option: -$OPTARG" >&2
 			;;
@@ -72,18 +83,39 @@ done
 DOC_MATRIX=${MATRIX_BUILDS}
 
 if [ -n "$MAIN_CI" ]; then
+
+	# West handling
+        pushd ..
+	if [ ! -d .west ]; then
+		west init -l zephyr
+		west update
+	fi
+        popd
+
 	if [ -z "$BRANCH" ]; then
 		echo "No base branch given"
-		exit
+		exit 1
 	else
 		COMMIT_RANGE=$REMOTE/${BRANCH}..HEAD
 		echo "Commit range:" ${COMMIT_RANGE}
 	fi
+	if [ -n "$RANGE" ]; then
+		COMMIT_RANGE=$RANGE
+	fi
 	source zephyr-env.sh
 	SANITYCHECK="${ZEPHYR_BASE}/scripts/sanitycheck"
+
+	# Possibly the only record of what exact version is being tested:
+	short_git_log='git log -n 5 --oneline --decorate --abbrev=12 '
+
 	if [ -n "$PULL_REQUEST_NR" ]; then
+		$short_git_log $REMOTE/${BRANCH}
+		# Now let's pray this script is being run from a
+		# different location
+# https://stackoverflow.com/questions/3398258/edit-shell-script-while-its-running
 		git rebase $REMOTE/${BRANCH};
 	fi
+	$short_git_log
 fi
 
 function handle_coverage() {
@@ -91,33 +123,40 @@ function handle_coverage() {
 	echo "Calling gcovr"
 	gcovr -r ${ZEPHYR_BASE} -x > shippable/codecoverage/coverage.xml;
 
-	# Capture data
-	echo "Running lcov --capture ..."
-	lcov --capture \
-		--directory sanity-out/native_posix/ \
-		--directory sanity-out/nrf52_bsim/ \
-		--directory sanity-out/unit_testing/ \
-		--directory bsim_bt_out/ \
-		--output-file lcov.pre.info -q --rc lcov_branch_coverage=1;
 
-	# Remove noise
-	echo "Exclude data from coverage report..."
-	lcov -q \
-		--remove lcov.pre.info mylib.c \
-		--remove lcov.pre.info tests/\* \
-		--remove lcov.pre.info samples/\* \
-		--remove lcov.pre.info ext/\* \
-		--remove lcov.pre.info *generated* \
-		-o lcov.info --rc lcov_branch_coverage=1;
+	# Upload to codecov.io only on merged builds or if CODECOV_IO variable
+	# is set.
+	if [ -n "${CODECOV_IO}" -o -z "${PULL_REQUEST_NR}" ]; then
+		# Capture data
+		echo "Running lcov --capture ..."
+		lcov --capture \
+			--directory sanity-out/native_posix/ \
+			--directory sanity-out/nrf52_bsim/ \
+			--directory sanity-out/unit_testing/ \
+			--directory bsim_bt_out/ \
+			--output-file lcov.pre.info -q --rc lcov_branch_coverage=1;
 
-	# Cleanup
-	rm lcov.pre.info;
+		# Remove noise
+		echo "Exclude data from coverage report..."
+		lcov -q \
+			--remove lcov.pre.info mylib.c \
+			--remove lcov.pre.info tests/\* \
+			--remove lcov.pre.info samples/\* \
+			--remove lcov.pre.info ext/\* \
+			--remove lcov.pre.info *generated* \
+			-o lcov.info --rc lcov_branch_coverage=1;
+
+		# Cleanup
+		rm lcov.pre.info;
+		rm -rf sanity-out out-2nd-pass;
+
+		# Upload to codecov.io
+		echo "Upload coverage reports to codecov.io"
+		bash <(curl -s https://codecov.io/bash) -f "lcov.info" -X coveragepy -X fixes;
+		rm -f lcov.info;
+	fi
+
 	rm -rf sanity-out out-2nd-pass;
-
-	# Upload to codecov.io
-	echo "Upload coverage reports to codecov.io"
-	bash <(curl -s https://codecov.io/bash) -f "lcov.info" -X coveragepy -X fixes;
-	rm -f lcov.info;
 
 }
 
@@ -152,6 +191,11 @@ function on_complete() {
 	if [ -e ${BSIM_BT_TEST_RESULTS_FILE} ]; then
 		echo "Copy ${BSIM_BT_TEST_RESULTS_FILE}"
 		cp ${BSIM_BT_TEST_RESULTS_FILE} shippable/testresults/;
+	fi;
+
+	if [ -e ${WEST_COMMANDS_RESULTS_FILE} ]; then
+		echo "Copy ${WEST_COMMANDS_RESULTS_FILE}"
+		cp ${WEST_COMMANDS_RESULTS_FILE} shippable/testresults;
 	fi;
 
 	if [ "$MATRIX" = "1" ]; then
@@ -192,17 +236,6 @@ function run_bsim_bt_tests() {
 	tests/bluetooth/bsim_bt/run_parallel.sh
 }
 
-function build_docs() {
-	echo "- Building Documentation";
-	make htmldocs || ( echo "Documentation build failed" ; exit 1 )
-
-	if [ -s doc/_build/doc.warnings ]; then
-		echo " => New documentation warnings/errors";
-		cp doc/_build/doc.warnings doc.warnings
-	fi
-	echo "- Verify commit message, coding style, doc build";
-}
-
 function get_tests_to_run() {
 	./scripts/ci/get_modified_tests.py --commits ${COMMIT_RANGE} > modified_tests.args;
 	./scripts/ci/get_modified_boards.py --commits ${COMMIT_RANGE} > modified_boards.args;
@@ -231,8 +264,23 @@ if [ -n "$MAIN_CI" ]; then
 		echo "Skipping BT simulator tests"
 	fi
 
+	if [ "$MATRIX" = "1" ]; then
+		# Run pytest-based testing for Python in matrix
+		# builder 1.  For now, this is just done for the west
+		# extension commands, but additional directories which
+		# run pytest could go here too.
+		PYTEST=$(type -p pytest-3 || echo "pytest")
+		mkdir -p $(dirname ${WEST_COMMANDS_RESULTS_FILE})
+		WEST_SRC=$(west list --format='{abspath}' west)/src
+		PYTHONPATH=./scripts/west_commands:$WEST_SRC "${PYTEST}" \
+			  --junitxml=${WEST_COMMANDS_RESULTS_FILE} \
+			  ./scripts/west_commands/tests
+	else
+		echo "Skipping west command tests"
+	fi
+
 	# In a pull-request see if we have changed any tests or board definitions
-	if [ -n "${PULL_REQUEST_NR}" ]; then
+	if [ -n "${PULL_REQUEST_NR}" -o -n "${LOCAL_RUN}"  ]; then
 		get_tests_to_run
 	fi
